# kv/RocksDBStore: Add minimum key limit before invoking DeleteRange.