@jdurgin 2019-11-07
How is the behavior with objects with >1024 keys?