@markhpc 2019-11-07
> credit @Aran85 in the commit message?

Yes absolutely!  I added a "co-authored-by" line with @aran85's email address.